requirements.txt
fastapi
uvicorn[standard]
websockets
opencv-python
mediapipe
ultralytics
numpy
motor
pydantic
python-dotenv